+++ spec/graphql/types/string_spec.rb
@@ -101,4 +101,48 @@ describe GraphQL::Types::String do
       assert_nil string_type.coerce_isolated_input(0.999)
     end
   end
+
+
+  describe "unicode escapes" do
+    class UnicodeEscapeSchema < GraphQL::Schema
+      class QueryType < GraphQL::Schema::Object
+        field :get_string, String do
+          argument :string, String
+        end
+
+        def get_string(string:)
+          string
+        end
+      end
+
+      query(QueryType)
+    end
+
+    it "parses escapes properly in single-quoted strings" do
+      query_str = File.read("./spec/fixtures/unicode_escapes/query1.graphql")
+      res = UnicodeEscapeSchema.execute(query_str)
+      assert_equal "d", res["data"]["example1"]
+      assert_equal "\\u0064", res["data"]["example2"]
+      assert_equal "\\u006", res["data"]["example4"]
+
+      error_query_str = query_str.gsub("  # ", "  ")
+      res2 = UnicodeEscapeSchema.execute(error_query_str)
+      expected_err = if USING_C_PARSER
+        "syntax error, unexpected invalid token (\"\\\"\") at [4, 31]"
+      else
+        "Expected string or block string, but it was malformed"
+      end
+      assert_equal [expected_err], res2["errors"].map { |err| err["message"] }
+    end
+
+    it "parses escapes properly in triple-quoted strings" do
+      query_str = File.read("./spec/fixtures/unicode_escapes/query2.graphql")
+      res = UnicodeEscapeSchema.execute(query_str)
+      # No replacing in block strings:
+      assert_equal "\\a", res["data"]["example1"]
+      assert_equal "\\u006", res["data"]["example2"]
+      assert_equal "\\n", res["data"]["example3"]
+      assert_equal "\\u0064", res["data"]["example4"]
+    end
+  end
 end